Move requirement validation to builder
